Huge commit of the day.
Haven't commited all day, very bad practice.

However I've gotten the AI to recognize the danger of a possible wumpus. It needs to be done towards pit aswell, not hard work. Gonna do it after this commit.

What needs to be fixed is when getting more information about the game world possible pits and wumpus need to be removed.